Simplify add/edit flight view
