+++ src/router.js
@@ -5,12 +5,6 @@ import Home from './views/Home'
 import Estimate from './views/Estimate'
 import Flights from './views/Flights'
 import AddEditFlight from './views/AddEditFlight'
-import AddEditFlightType from './views/AddFlightType'
-import AddEditFlightDeparture from './views/AddEditFlightDeparture'
-import AddEditFlightArrival from './views/AddEditFlightArrival'
-import AddEditFlightNumberForm from './views/AddEditFlightNumberForm'
-// import AddEditFlightDate from './views/AddEditFlightDate'
-import AddEditFlightPassengers from './views/AddEditFlightPassengers'
 import Checkout from './views/Checkout'
 import Success from './views/Success'
 import About from './views/About'
@@ -18,28 +12,6 @@ import Privacy from './views/Privacy'
 import NotFound from './views/NotFound'
 
 Vue.use(VueRouter)
-
-const getFlightFormChildren = namePrefix => [{
-  name: `${namePrefix}-departure`,
-  path: 'departure',
-  component: AddEditFlightDeparture,
-  props: true
-}, {
-  name: `${namePrefix}-arrival`,
-  path: 'arrival',
-  component: AddEditFlightArrival,
-  props: true
-}, {
-  name: `${namePrefix}-number`,
-  path: 'flight-number',
-  component: AddEditFlightNumberForm,
-  props: true
-}, {
-  name: `${namePrefix}-passengers`,
-  path: 'passengers',
-  component: AddEditFlightPassengers,
-  props: true
-}]
 
 const routes = [
   {
@@ -58,22 +30,12 @@ const routes = [
     }, {
       name: 'add-flight',
       path: 'add-flight',
-      component: AddEditFlight,
-      children: [
-        {
-          name: 'add-flight-type',
-          path: 'type',
-          component: AddEditFlightType,
-          props: true
-        },
-        ...getFlightFormChildren('add-flight')
-      ]
+      component: AddEditFlight
     }, {
       name: 'edit-flight',
       path: 'edit-flight/:id',
       component: AddEditFlight,
-      props: true,
-      children: getFlightFormChildren('edit-flight')
+      props: true
     }, {
       name: 'checkout',
       path: 'checkout',
